NTP-969: Initial commit with slight restructure to enquiry pages, so starts from search and has guidance and confirm pages
