+++ Application/Common/DTO/EnquirerEnquiryResponseReceivedDto.cs
@@ -2,7 +2,7 @@ using Application.Common.Models;
 
 namespace Application.Common.DTO;
 
-public class EnquirerEnquiryResponseReceivedDto : EnquiryResponseModel
+public record EnquirerEnquiryResponseReceivedDto : EnquiryResponseModel
 {
     public string TuitionPartnerName { get; set; } = string.Empty;
 }
