+++ Application/Common/Models/EnquiryModel.cs
@@ -1,9 +1,12 @@
+using Domain.Search;
+
 namespace Application.Common.Models;
 
-public class EnquiryModel : EnquiryBaseModel
+public record EnquiryModel : EnquiryBaseModel
 {
     public string? EnquiryText { get; set; }
 
     public string? Email { get; set; }
-    public List<string>? SelectedTuitionPartners { get; set; }
+
+    public TuitionPartnersResult? TuitionPartnersForEnquiry { get; set; }
 }
