+++ UI/Pages/Index.cshtml
@@ -46,6 +46,7 @@
             <input asp-for="Data.CompareListOrderByDirection" type="hidden" />
             <input asp-for="Data.CompareListTuitionType" type="hidden" />
             <input asp-for="Data.CompareListGroupSize" type="hidden" />
+            <input asp-for="Data.CompareListShowWithVAT" type="hidden" />
 
             <h2 class="govuk-heading-m" data-testid="other-tutoring-options">Other tutoring options</h2>
             <p class="govuk-body">You can use your National Tutoring Programme funding to mix and match tutoring options.</p>
