+++ UI/Pages/WhichSubjects.cshtml
@@ -47,6 +47,7 @@
             <input asp-for="Data.CompareListOrderByDirection" type="hidden" />
             <input asp-for="Data.CompareListTuitionType" type="hidden" />
             <input asp-for="Data.CompareListGroupSize" type="hidden" />
+            <input asp-for="Data.CompareListShowWithVAT" type="hidden" />
 
             <div class="govuk-button-group">
                 <govuk-button type="submit" data-testid="call-to-action">Continue</govuk-button>
